Progressbar Testing, cleaning up code
